:lipstick: Select Kernel command impl in notebook core instead of contrib.
