+++ src/vs/workbench/contrib/notebook/browser/contrib/editorStatusBar/editorStatusBar.ts
@@ -3,492 +3,23 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { groupBy } from 'vs/base/common/arrays';
-import { createCancelablePromise } from 'vs/base/common/async';
-import { CancellationToken } from 'vs/base/common/cancellation';
-import { Codicon } from 'vs/base/common/codicons';
-import { Event } from 'vs/base/common/event';
 import { Disposable, DisposableStore, IDisposable, MutableDisposable } from 'vs/base/common/lifecycle';
 import { Schemas } from 'vs/base/common/network';
-import { compareIgnoreCase, uppercaseFirstLetter } from 'vs/base/common/strings';
 import { ILanguageFeaturesService } from 'vs/editor/common/services/languageFeatures';
 import * as nls from 'vs/nls';
-import { Action2, MenuId, registerAction2 } from 'vs/platform/actions/common/actions';
-import { ContextKeyExpr, IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
-import { ExtensionIdentifier } from 'vs/platform/extensions/common/extensions';
-import { IInstantiationService, ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
-import { ILabelService } from 'vs/platform/label/common/label';
+import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { ILogService } from 'vs/platform/log/common/log';
-import { IProductService } from 'vs/platform/product/common/productService';
-import { ProgressLocation } from 'vs/platform/progress/common/progress';
-import { IQuickInputService, IQuickPick, IQuickPickItem, QuickPickInput } from 'vs/platform/quickinput/common/quickInput';
 import { Registry } from 'vs/platform/registry/common/platform';
 import { Extensions as WorkbenchExtensions, IWorkbenchContribution, IWorkbenchContributionsRegistry } from 'vs/workbench/common/contributions';
-import { ViewContainerLocation } from 'vs/workbench/common/views';
-import { IExtensionsViewPaneContainer, IExtensionsWorkbenchService, VIEWLET_ID as EXTENSION_VIEWLET_ID } from 'vs/workbench/contrib/extensions/common/extensions';
 import { CENTER_ACTIVE_CELL } from 'vs/workbench/contrib/notebook/browser/contrib/navigation/arrow';
-import { NOTEBOOK_ACTIONS_CATEGORY, SELECT_KERNEL_ID } from 'vs/workbench/contrib/notebook/browser/controller/coreActions';
-import { getNotebookEditorFromEditorPane, INotebookEditor, INotebookExtensionRecommendation, KERNEL_RECOMMENDATIONS } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
-import { NotebookEditorWidget } from 'vs/workbench/contrib/notebook/browser/notebookEditorWidget';
-import { selectKernelIcon } from 'vs/workbench/contrib/notebook/browser/notebookIcons';
+import { SELECT_KERNEL_ID } from 'vs/workbench/contrib/notebook/browser/controller/coreActions';
+import { getNotebookEditorFromEditorPane, INotebookEditor } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
 import { NotebookTextModel } from 'vs/workbench/contrib/notebook/common/model/notebookTextModel';
 import { NotebookCellsChangeType } from 'vs/workbench/contrib/notebook/common/notebookCommon';
-import { NOTEBOOK_IS_ACTIVE_EDITOR, NOTEBOOK_KERNEL_COUNT } from 'vs/workbench/contrib/notebook/common/notebookContextKeys';
-import { INotebookKernel, INotebookKernelMatchResult, INotebookKernelService, ISourceAction } from 'vs/workbench/contrib/notebook/common/notebookKernelService';
+import { INotebookKernel, INotebookKernelService } from 'vs/workbench/contrib/notebook/common/notebookKernelService';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
-import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 import { LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle';
-import { IPaneCompositePartService } from 'vs/workbench/services/panecomposite/browser/panecomposite';
 import { IStatusbarEntry, IStatusbarEntryAccessor, IStatusbarService, StatusbarAlignment } from 'vs/workbench/services/statusbar/browser/statusbar';
-
-type KernelPick = IQuickPickItem & { kernel: INotebookKernel };
-function isKernelPick(item: QuickPickInput<IQuickPickItem>): item is KernelPick {
-	return 'kernel' in item;
-}
-type SourcePick = IQuickPickItem & { action: ISourceAction };
-function isSourcePick(item: QuickPickInput<IQuickPickItem>): item is SourcePick {
-	return 'action' in item;
-}
-type InstallExtensionPick = IQuickPickItem & { extensionId: string };
-function isInstallExtensionPick(item: QuickPickInput<IQuickPickItem>): item is InstallExtensionPick {
-	return item.id === 'installSuggested' && 'extensionId' in item;
-}
-type KernelQuickPickItem = IQuickPickItem | InstallExtensionPick | KernelPick | SourcePick;
-const KERNEL_PICKER_UPDATE_DEBOUNCE = 200;
-
-registerAction2(class extends Action2 {
-	constructor() {
-		super({
-			id: SELECT_KERNEL_ID,
-			category: NOTEBOOK_ACTIONS_CATEGORY,
-			title: { value: nls.localize('notebookActions.selectKernel', "Select Notebook Kernel"), original: 'Select Notebook Kernel' },
-			// precondition: NOTEBOOK_IS_ACTIVE_EDITOR,
-			icon: selectKernelIcon,
-			f1: true,
-			menu: [{
-				id: MenuId.EditorTitle,
-				when: ContextKeyExpr.and(
-					NOTEBOOK_IS_ACTIVE_EDITOR,
-					ContextKeyExpr.notEquals('config.notebook.globalToolbar', true)
-				),
-				group: 'navigation',
-				order: -10
-			}, {
-				id: MenuId.NotebookToolbar,
-				when: ContextKeyExpr.equals('config.notebook.globalToolbar', true),
-				group: 'status',
-				order: -10
-			}, {
-				id: MenuId.InteractiveToolbar,
-				when: NOTEBOOK_KERNEL_COUNT.notEqualsTo(0),
-				group: 'status',
-				order: -10
-			}],
-			description: {
-				description: nls.localize('notebookActions.selectKernel.args', "Notebook Kernel Args"),
-				args: [
-					{
-						name: 'kernelInfo',
-						description: 'The kernel info',
-						schema: {
-							'type': 'object',
-							'required': ['id', 'extension'],
-							'properties': {
-								'id': {
-									'type': 'string'
-								},
-								'extension': {
-									'type': 'string'
-								},
-								'notebookEditorId': {
-									'type': 'string'
-								}
-							}
-						}
-					}
-				]
-			},
-		});
-	}
-
-	async run(accessor: ServicesAccessor, context?:
-		{ id: string; extension: string } |
-		{ notebookEditorId: string } |
-		{ id: string; extension: string; notebookEditorId: string } |
-		{ ui?: boolean; notebookEditor?: NotebookEditorWidget } |
-		undefined
-	): Promise<boolean> {
-		const notebookKernelService = accessor.get(INotebookKernelService);
-		const editorService = accessor.get(IEditorService);
-		const productService = accessor.get(IProductService);
-		const quickInputService = accessor.get(IQuickInputService);
-		const labelService = accessor.get(ILabelService);
-		const logService = accessor.get(ILogService);
-		const paneCompositeService = accessor.get(IPaneCompositePartService);
-		const extensionWorkbenchService = accessor.get(IExtensionsWorkbenchService);
-		const extensionHostService = accessor.get(IExtensionService);
-
-		let editor: INotebookEditor | undefined;
-		if (context !== undefined && 'notebookEditorId' in context) {
-			const editorId = context.notebookEditorId;
-			const matchingEditor = editorService.visibleEditorPanes.find((editorPane) => {
-				const notebookEditor = getNotebookEditorFromEditorPane(editorPane);
-				return notebookEditor?.getId() === editorId;
-			});
-			editor = getNotebookEditorFromEditorPane(matchingEditor);
-		} else if (context !== undefined && 'notebookEditor' in context) {
-			editor = context?.notebookEditor;
-		} else {
-			editor = getNotebookEditorFromEditorPane(editorService.activeEditorPane);
-		}
-
-		if (!editor || !editor.hasModel()) {
-			return false;
-		}
-		let controllerId = context && 'id' in context ? context.id : undefined;
-		let extensionId = context && 'extension' in context ? context.extension : undefined;
-
-		if (controllerId && (typeof controllerId !== 'string' || typeof extensionId !== 'string')) {
-			// validate context: id & extension MUST be strings
-			controllerId = undefined;
-			extensionId = undefined;
-		}
-
-		const notebook = editor.textModel;
-		const scopedContextKeyService = editor.scopedContextKeyService;
-		const matchResult = notebookKernelService.getMatchingKernel(notebook);
-		const { selected, all } = matchResult;
-
-		if (selected && controllerId && selected.id === controllerId && ExtensionIdentifier.equals(selected.extension, extensionId)) {
-			// current kernel is wanted kernel -> done
-			return true;
-		}
-
-		let newKernel: INotebookKernel | undefined;
-		if (controllerId) {
-			const wantedId = `${extensionId}/${controllerId}`;
-			for (const candidate of all) {
-				if (candidate.id === wantedId) {
-					newKernel = candidate;
-					break;
-				}
-			}
-			if (!newKernel) {
-				logService.warn(`wanted kernel DOES NOT EXIST, wanted: ${wantedId}, all: ${all.map(k => k.id)}`);
-				return false;
-			}
-		}
-
-		if (newKernel) {
-			notebookKernelService.selectKernelForNotebook(newKernel, notebook);
-			return true;
-		}
-
-		const quickPick = quickInputService.createQuickPick<KernelQuickPickItem>();
-		const quickPickItems = this._getKernelPickerQuickPickItems(notebook, matchResult, notebookKernelService, scopedContextKeyService);
-		quickPick.items = quickPickItems;
-		quickPick.canSelectMany = false;
-		quickPick.placeholder = selected
-			? nls.localize('prompt.placeholder.change', "Change kernel for '{0}'", labelService.getUriLabel(notebook.uri, { relative: true }))
-			: nls.localize('prompt.placeholder.select', "Select kernel for '{0}'", labelService.getUriLabel(notebook.uri, { relative: true }));
-
-		quickPick.busy = notebookKernelService.getKernelDetectionTasks(notebook).length > 0;
-
-		const kernelDetectionTaskListener = notebookKernelService.onDidChangeKernelDetectionTasks(() => {
-			quickPick.busy = notebookKernelService.getKernelDetectionTasks(notebook).length > 0;
-		});
-
-		// run extension recommendataion task if quickPickItems is empty
-		const extensionRecommendataionPromise = quickPickItems.length === 0
-			? createCancelablePromise(token => this._showInstallKernelExtensionRecommendation(notebook, quickPick, extensionWorkbenchService, token))
-			: undefined;
-
-		const kernelChangeEventListener = Event.debounce<void, void>(
-			Event.any(
-				notebookKernelService.onDidChangeSourceActions,
-				notebookKernelService.onDidAddKernel,
-				notebookKernelService.onDidRemoveKernel,
-				notebookKernelService.onDidChangeNotebookAffinity
-			),
-			(last, _current) => last,
-			KERNEL_PICKER_UPDATE_DEBOUNCE
-		)(async () => {
-			// reset quick pick progress
-			quickPick.busy = false;
-			extensionRecommendataionPromise?.cancel();
-
-			const currentActiveItems = quickPick.activeItems;
-			const matchResult = notebookKernelService.getMatchingKernel(notebook);
-			const quickPickItems = this._getKernelPickerQuickPickItems(notebook, matchResult, notebookKernelService, scopedContextKeyService);
-			quickPick.keepScrollPosition = true;
-
-			// recalcuate active items
-			const activeItems: KernelQuickPickItem[] = [];
-			for (const item of currentActiveItems) {
-				if (isKernelPick(item)) {
-					const kernelId = item.kernel.id;
-					const sameItem = quickPickItems.find(pi => isKernelPick(pi) && pi.kernel.id === kernelId) as KernelPick | undefined;
-					if (sameItem) {
-						activeItems.push(sameItem);
-					}
-				} else if (isSourcePick(item)) {
-					const sameItem = quickPickItems.find(pi => isSourcePick(pi) && pi.action.action.id === item.action.action.id) as SourcePick | undefined;
-					if (sameItem) {
-						activeItems.push(sameItem);
-					}
-				}
-			}
-
-			quickPick.items = quickPickItems;
-			quickPick.activeItems = activeItems;
-		}, this);
-
-		const pick = await new Promise<KernelQuickPickItem>((resolve, reject) => {
-			quickPick.onDidAccept(() => {
-				const item = quickPick.selectedItems[0];
-				if (item) {
-					resolve(item);
-				} else {
-					reject();
-				}
-
-				quickPick.hide();
-			});
-
-			quickPick.onDidHide(() => () => {
-				kernelDetectionTaskListener.dispose();
-				kernelChangeEventListener.dispose();
-				quickPick.dispose();
-				reject();
-			});
-			quickPick.show();
-		});
-
-		if (pick) {
-			if (isKernelPick(pick)) {
-				newKernel = pick.kernel;
-				notebookKernelService.selectKernelForNotebook(newKernel, notebook);
-				return true;
-			}
-
-			// actions
-			if (pick.id === 'install') {
-				await this._showKernelExtension(
-					paneCompositeService,
-					extensionWorkbenchService,
-					extensionHostService,
-					notebook.viewType
-				);
-				// suggestedExtension must be defined for this option to be shown, but still check to make TS happy
-			} else if (isInstallExtensionPick(pick)) {
-				await this._showKernelExtension(
-					paneCompositeService,
-					extensionWorkbenchService,
-					extensionHostService,
-					notebook.viewType,
-					pick.extensionId,
-					productService.quality !== 'stable'
-				);
-			} else if (isSourcePick(pick)) {
-				// selected explicilty, it should trigger the execution?
-				pick.action.runAction();
-			}
-		}
-
-		return false;
-	}
-
-	private _getKernelPickerQuickPickItems(
-		notebookTextModel: NotebookTextModel,
-		matchResult: INotebookKernelMatchResult,
-		notebookKernelService: INotebookKernelService,
-		scopedContextKeyService: IContextKeyService
-	): QuickPickInput<KernelQuickPickItem>[] {
-		const { selected, all, suggestions, hidden } = matchResult;
-
-		function toQuickPick(kernel: INotebookKernel) {
-			const res = <KernelPick>{
-				kernel,
-				picked: kernel.id === selected?.id,
-				label: kernel.label,
-				description: kernel.description,
-				detail: kernel.detail
-			};
-			if (kernel.id === selected?.id) {
-				if (!res.description) {
-					res.description = nls.localize('current1', "Currently Selected");
-				} else {
-					res.description = nls.localize('current2', "{0} - Currently Selected", res.description);
-				}
-			}
-			return res;
-		}
-		const quickPickItems: QuickPickInput<KernelQuickPickItem>[] = [];
-		if (all.length) {
-			// Always display suggested kernels on the top.
-			if (suggestions.length) {
-				quickPickItems.push({
-					type: 'separator',
-					label: nls.localize('suggestedKernels', "Suggested")
-				});
-				quickPickItems.push(...suggestions.map(toQuickPick));
-			}
-
-			// Next display all of the kernels not marked as hidden grouped by categories or extensions.
-			// If we don't have a kind, always display those at the bottom.
-			const picks = all.filter(item => (!suggestions.includes(item) && !hidden.includes(item))).map(toQuickPick);
-			const kernelsPerCategory = groupBy(picks, (a, b) => compareIgnoreCase(a.kernel.kind || 'z', b.kernel.kind || 'z'));
-			kernelsPerCategory.forEach(items => {
-				quickPickItems.push({
-					type: 'separator',
-					label: items[0].kernel.kind || nls.localize('otherKernelKinds', "Other")
-				});
-				quickPickItems.push(...items);
-			});
-		}
-
-		const sourceActions = notebookKernelService.getSourceActions(notebookTextModel, scopedContextKeyService);
-		if (sourceActions.length) {
-			quickPickItems.push({
-				type: 'separator',
-				// label: nls.localize('sourceActions', "")
-			});
-
-			sourceActions.forEach(sourceAction => {
-				const res = <SourcePick>{
-					action: sourceAction,
-					picked: false,
-					label: sourceAction.action.label,
-				};
-
-				quickPickItems.push(res);
-			});
-		}
-
-		return quickPickItems;
-	}
-
-	private async _showInstallKernelExtensionRecommendation(
-		notebookTextModel: NotebookTextModel,
-		quickPick: IQuickPick<KernelQuickPickItem>,
-		extensionWorkbenchService: IExtensionsWorkbenchService,
-		token: CancellationToken
-	) {
-		quickPick.busy = true;
-
-		const newQuickPickItems = await this._getKernelRecommendationsQuickPickItems(notebookTextModel, extensionWorkbenchService);
-		quickPick.busy = false;
-
-		if (token.isCancellationRequested) {
-			return;
-		}
-
-		if (newQuickPickItems && quickPick.items.length === 0) {
-			quickPick.items = newQuickPickItems;
-		}
-	}
-
-
-	private async _getKernelRecommendationsQuickPickItems(
-		notebookTextModel: NotebookTextModel,
-		extensionWorkbenchService: IExtensionsWorkbenchService,
-	): Promise<QuickPickInput<KernelQuickPickItem>[] | undefined> {
-		const quickPickItems: QuickPickInput<KernelQuickPickItem>[] = [];
-
-		const language = this.getSuggestedLanguage(notebookTextModel);
-		const suggestedExtension: INotebookExtensionRecommendation | undefined = language ? this.getSuggestedKernelFromLanguage(notebookTextModel.viewType, language) : undefined;
-		if (suggestedExtension) {
-			await extensionWorkbenchService.queryLocal();
-			const extension = extensionWorkbenchService.installed.find(e => e.identifier.id === suggestedExtension.extensionId);
-
-			if (extension) {
-				// it's installed but might be detecting kernels
-				return undefined;
-			}
-
-			// We have a suggested kernel, show an option to install it
-			quickPickItems.push({
-				id: 'installSuggested',
-				description: suggestedExtension.displayName ?? suggestedExtension.extensionId,
-				label: `$(${Codicon.lightbulb.id}) ` + nls.localize('installSuggestedKernel', 'Install suggested extensions'),
-				extensionId: suggestedExtension.extensionId
-			});
-		}
-		// there is no kernel, show the install from marketplace
-		quickPickItems.push({
-			id: 'install',
-			label: nls.localize('searchForKernels', "Browse marketplace for kernel extensions"),
-		});
-
-		return quickPickItems;
-	}
-
-	/**
-	 * Examine the most common language in the notebook
-	 * @param notebookTextModel The notebook text model
-	 * @returns What the suggested language is for the notebook. Used for kernal installing
-	 */
-	private getSuggestedLanguage(notebookTextModel: NotebookTextModel): string | undefined {
-		const metaData = notebookTextModel.metadata;
-		let suggestedKernelLanguage: string | undefined = (metaData.custom as any)?.metadata?.language_info?.name;
-		// TODO how do we suggest multi language notebooks?
-		if (!suggestedKernelLanguage) {
-			const cellLanguages = notebookTextModel.cells.map(cell => cell.language).filter(language => language !== 'markdown');
-			// Check if cell languages is all the same
-			if (cellLanguages.length > 1) {
-				const firstLanguage = cellLanguages[0];
-				if (cellLanguages.every(language => language === firstLanguage)) {
-					suggestedKernelLanguage = firstLanguage;
-				}
-			}
-		}
-		return suggestedKernelLanguage;
-	}
-
-	/**
-	 * Given a language and notebook view type suggest a kernel for installation
-	 * @param language The language to find a suggested kernel extension for
-	 * @returns A recommednation object for the recommended extension, else undefined
-	 */
-	private getSuggestedKernelFromLanguage(viewType: string, language: string): INotebookExtensionRecommendation | undefined {
-		const recommendation = KERNEL_RECOMMENDATIONS.get(viewType)?.get(language);
-		return recommendation;
-	}
-
-	private async _showKernelExtension(
-		paneCompositePartService: IPaneCompositePartService,
-		extensionWorkbenchService: IExtensionsWorkbenchService,
-		extensionService: IExtensionService,
-		viewType: string,
-		extId?: string,
-		isInsiders?: boolean
-	) {
-		// If extension id is provided attempt to install the extension as the user has requested the suggested ones be installed
-		if (extId) {
-			const extension = (await extensionWorkbenchService.getExtensions([{ id: extId }], CancellationToken.None))[0];
-			const canInstall = await extensionWorkbenchService.canInstall(extension);
-			// If we can install then install it, otherwise we will fall out into searching the viewlet
-			if (canInstall) {
-				await extensionWorkbenchService.install(
-					extension,
-					{
-						installPreReleaseVersion: isInsiders ?? false,
-						context: { skipWalkthrough: true }
-					},
-					ProgressLocation.Notification
-				);
-				await extensionService.activateByEvent(`onNotebook:${viewType}`);
-				return;
-			}
-		}
-
-		const viewlet = await paneCompositePartService.openPaneComposite(EXTENSION_VIEWLET_ID, ViewContainerLocation.Sidebar, true);
-		const view = viewlet?.getViewPaneContainer() as IExtensionsViewPaneContainer | undefined;
-		const pascalCased = viewType.split(/[^a-z0-9]/ig).map(uppercaseFirstLetter).join('');
-		view?.search(`@tag:notebookKernel${pascalCased}`);
-	}
-});
-
 
 class ImplictKernelSelector implements IDisposable {
